+++ test/items/pn-model/__check__/ts/model.ts
@@ -1,0 +1,25 @@
+export namespace PN_Mandor {
+  /**
+  * 
+  */
+  export const thisName: string = "Mandor";
+  /**
+   * Required. 
+   */
+  export const prop0: string = "prop0";
+  /**
+   * Required. 
+   */
+  export const prop1: string = "prop1";
+}
+
+export interface Mandor {
+  /**
+   * Required. 
+   */
+  prop0: string;
+  /**
+   * Required. 
+   */
+  prop1: number;
+}